chore: move 'utils/rules.ts' to 'rules/index.ts'
- The `lib/load-rules.ts` file is confusing because it behaves like an index file for the src/rules folder, so adding a comment helps clarify its purpose.

- Rename `tools/update-rules.ts` to `tools/update-rules-index.ts`.
- Move `utils/rules.ts` to `rules/index.ts`
